rfid_operations_api/requirements.txt
# RFID Operations API - Python Dependencies

# Core FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration and environment
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0

# HTTP and API utilities
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Data validation and serialization
email-validator==2.1.0
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# Production server
gunicorn==21.2.0

# Monitoring and logging
structlog==23.2.0

# RFID/QR code support (for future integration)
opencv-python==4.8.1.78
pillow==10.1.0

# Tailscale/network utilities
netifaces==0.11.0